More public API cleanups.
- scalafix.Failure is moved to scalafix.internal.Failure, this was only
  used by pass information between internal code and the cli.
- RewriteCtx.debug, as convenient as this method was I think
  users should use pprint.log instead. The method had nothing
  to do with rewriting.
- scalafix.config is removed entirely, now that we have Rewrite.init
  there is no need for users to get an instance of ScalafixConfig.
- unused scalafix.util.TreeExtractors.
